feat: Describe support for audio uploads as binary
Describe how to support audio upload endpoints expecting to have a binary in the POST request body instead of forms.
By default, the Demo Web App expects the endpoint to handle multipart forms unless you set `FILE_UPLOADING_FORMAT` to `binary`, which is the case for the example implementation of the .Net 7 Isolated function.

noissue